Move profile links array to its own file; import it to header and footer files to map links dynamically.
