+++ src/Components/ProfileLinksArray.jsx
@@ -1,0 +1,22 @@
+const profileLinksArray = [
+    {
+      href: 'https://github.com/joshleewillis',
+      src: './assets/Logo_Images/github-logo.png',
+      alt: 'Github Logo',
+      title: 'Github'
+    },
+    {
+      href: 'https://www.linkedin.com/in/joshua-willis-45794293/',
+      src: './assets/Logo_Images/linkedin-logo.png',
+      alt: 'Linkedin Logo',
+      title: 'Linkedin'
+    },
+    {
+      href: 'https://www.freecodecamp.org/joshleewillis',
+      src: './assets/Logo_Images/free-code-camp-logo.png',
+      alt: 'FreeCodeCamp Logo',
+      title: 'FreeCodeCamp'
+    }
+  ];
+
+  export default profileLinksArray;
